# public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BetAki Admin API",
        "description": "API REST da Dashboard de Administração da Betaki (Laravel 12 / PHP 8.4).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Ambiente atual"
        }
    ],
    "paths": {
...
        "/api/v1/slots": {
            "get": {
                "tags": [
                    "Slots"
                ],
                "summary": "Listar slots",
                "operationId": "7617b91dea798b9b0535615e58f38204",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Slots"
                ],
                "summary": "Criar slot",
                "operationId": "e024679ff643314058085e14ca635a11",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Criado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/slots/{id}": {
            "get": {
                "tags": [
                    "Slots"
                ],
                "summary": "Detalhar slot",
                "operationId": "fc7b293dd85066629891b53a62023464",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Slots"
                ],
                "summary": "Atualizar slot",
                "operationId": "4d13e7bc130eb9ac31d773b506ddc6da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Slots"
                ],
                "summary": "Remover slot",
                "operationId": "b446f6eb825dcbd3fb20be6d05efde9f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sem conteúdo"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/winners/batches": {
            "get": {
                "tags": [
                    "TopWinners"
                ],
                "summary": "Listar lotes (Top Winners)",
                "operationId": "e5b58a51c5e9d2f4b4eb4c412731837e",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "TopWinners"
                ],
                "summary": "Criar lote",
                "operationId": "2db42d2f26cbe76803158caa07c3f8f3",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Criado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/winners/batches/{batch}": {
            "get": {
                "tags": [
                    "TopWinners"
                ],
                "summary": "Detalhar lote + vencedores",
                "operationId": "7edf156a83f8df93d95f3eb0431fd783",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "TopWinners"
                ],
                "summary": "Atualizar lote",
                "operationId": "e7cbb217f6a287b1b81e1c8b4c49fa71",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "TopWinners"
                ],
                "summary": "Remover lote",
                "operationId": "637c939278828d4dda947b27e52d3ad8",
                "responses": {
                    "204": {
                        "description": "Sem conteúdo"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/winners/batches/{batch}/results": {
            "put": {
                "tags": [
                    "TopWinners"
                ],
                "summary": "Sincronizar vencedores (pós-apuração) para revisão",
                "operationId": "6c900b8d264585c3659bf571deb0f263",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/winners/batches/{batch}/publish": {
            "post": {
                "tags": [
                    "TopWinners"
                ],
                "summary": "Publicar lote revisado",
                "operationId": "9670319c541aa1911f0cde5f34a5828f",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/winners/batches/{batch}/archive": {
            "post": {
                "tags": [
                    "TopWinners"
                ],
                "summary": "Arquivar lote publicado",
                "operationId": "eef154922e3973745fe0ae0f8bb3e07c",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Envie: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}